Add files via upload
Instructions
Create a smart contract that saves user balance. The contract should have the functions:

deposit (uint256 amount) this function accepts one argument and it saves the amount a user is depositing into a mapping,
checkBalance() this function searches for the user balance inside the balance mapping and returns the balance of whoever is calling the contract (i.e msg.sender). This function does not accept any arguments.