Add files via upload
This task extends the functionality of ThirdAssignment. I created a deposit function that allows anybody to send funds. The user and the amount will be stored in a mapping as the previous task.

I added a withdraw function and created a modifier that only allows the owner of the contract to withdraw the funds.
I also added the addFund function and created a modifier that only allows users that have deposited using the deposit function, to increase their balance on the mapping. The function accepts the amount to be added and updates the mapping to have the new balance. 
Finally, I created a modifier called feeCovered that checks that the additional funds are more than the set Fee before the addFund function is executed.